Kernel/USBMS: Move the BOT/BBB driver to its own directory
